string tests successfully added
